requirements.txt
mock==2.0.0
nose==1.3.7
PyYAML==3.12
requests==2.18.4
